+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-# Prereleased-2025.10.14
-- [Gelbooru 0.2] Add favorites and login support, you will need to provide additional information in the profile for it to work
+# Prereleased-2025.11.02
+- [Gelbooru 0.2, Shimmie2] Add login and favorite support
 - Add random sort option for bookmark
 - Add a button to lock screen when viewing video, also keep screen awake during video playback
 - Double tap on the left/right side of the screen to seek backward/forward in video player